requirements.txt
azure-cognitiveservices-vision-face
dlib
numpy
Pillow
opencv-python
